move Slot logic
